+++ src/reader/geometry_collection.rs
@@ -1,6 +1,6 @@
 use std::io::Cursor;
 
-use crate::common::WkbDimension;
+use crate::common::Dimension;
 use crate::error::WkbResult;
 use crate::reader::util::{has_srid, ReadBytesExt};
 use crate::reader::Wkb;
@@ -15,12 +15,12 @@ const HEADER_BYTES: u64 = 5;
 pub struct GeometryCollection<'a> {
     /// A WKB object for each of the internal geometries
     geometries: Vec<Wkb<'a>>,
-    dim: WkbDimension,
+    dim: Dimension,
     has_srid: bool,
 }
 
 impl<'a> GeometryCollection<'a> {
-    pub fn try_new(buf: &'a [u8], byte_order: Endianness, dim: WkbDimension) -> WkbResult<Self> {
+    pub fn try_new(buf: &'a [u8], byte_order: Endianness, dim: Dimension) -> WkbResult<Self> {
         let mut offset = 0;
         let has_srid = has_srid(buf, byte_order, offset);
         if has_srid {
@@ -53,7 +53,7 @@ impl<'a> GeometryCollection<'a> {
         })
     }
 
-    pub fn dimension(&self) -> WkbDimension {
+    pub fn dimension(&self) -> Dimension {
         self.dim
     }